requirements.txt
streamlit>=1.37.0,<2.0
pandas>=2.2.0,<3.0
requests>=2.32.0,<3.0
beautifulsoup4>=4.12.2,<5.0
openai>=1.30.0,<2.0
python-dotenv>=1.0.0,<2.0
google-api-python-client>=2.136.0,<3.0
